adds some more error handling

+++ src/js/api/posts/delete.mjs
@@ -9,6 +9,10 @@ const method = "delete";
 export async function removePost(id) {
     const updatePostUrl = apiSocialUrl + action + "/" + id;
 
+    if (!id) {
+        throw new Error ("ID not defined. Delete requires a post ID")
+    }
+
     const response = await authFetch(updatePostUrl, {
         method: "delete"
     })
